requirements.txt
#arrange in alphabetic order
#apache-airflow[postgres, celery, rabbitmq]
calamari_ocr==0.3.3
django
djangorestframework
drive-cli
flask
flask_cors
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gin-config
grpcio
matplotlib
numpy
pandas
py4j==0.10.7
pytesseract
psycopg2
requests
scipy
sklearn
shapely
spacy
tensorflow
tensorflow-gpu
tensorflow-serving-api
tensorflow-hub
tqdm
overrides
opencv-python
imutils
keras
xmltodict
torch=1.2.0
fire
lmdb
torchvision
natsort
fire
nltk






